+++ CHANGES.md
@@ -40,4 +40,5 @@
 * updated Ubuntu deployment scripts
 * derive pub key from cert; priv key encrypted with user password, not available on smtp relay
 * multiple ports connection option for SMTP server
-* added simple library and utility for stun protocol
+* added simple library and utility for stun protocol to obtain mapped address of the server
+* added Received: header to relayed email; includes private and mapped address; could be used for direct communication between peers